Enable magic at start of syntax regexes
The '\v' enables "magic" for the regex that follows.
Thus, it should be written at the beginning of the regex, not after the
"start of line" special character.

Technically, it's not needed at all, as all `syntax match` regexes are
always interpreted as having the "magic" flag set, but this makes it
explicit, and thus easier to keep in mind.